+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+### Changed
+- require Cython 3.3+ so free-threaded Python 3.14 can compile
+- mark the extension free-threading compatible (one SequenceMatcher per thread;
+  HtmlDiff's shared prefix counter is locked)
+- add support for Python 3.14 and 3.14t
+- allow CMake 3.15 through 3.30
+
 ## [1.2.0] - 2025-04-11
 ### Changed
 - drop support for Python 3.8